+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 
 ### 0.13.2-dev
 
-
+* `lobster-codebeamer`
+  - The retry logic uses exponential backoff feature instead of retrying successively
+    without any time gap. The exponential backoff factor is set to 1 so the 
+    interval of calls will be 1s, 2s, 4s, 8s and so on.
 
 ### 0.13.1